test/fixtures/wpt/resource-timing/resource_dedicated_worker.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing in dedicated workers</title>
<link rel="author" title="Google" href="http://www.google.com/" />
<link rel="help" href="http://www.w3.org/TR/resource-timing/"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/webperftestharness.js"></script>
<script src="resources/webperftestharnessextension.js"></script>
<link rel="stylesheet" href="resources/resource_timing_test0.css" />
<script>
setup({explicit_done: true});
const worker = new Worker("resources/worker_with_images.js");
worker.onmessage = function(event) {
  const context = new PerformanceContext(window.performance);
  const entries = context.getEntriesByType('resource');
  test_equals(entries.length, 6, "There should be six entries: 4 scripts, 1 stylesheet, and the worker itself");
  done();
}
</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that resources requested by dedicated workers don't appear in the main document.</p>
</body>
</html>
