src/third_party/blink/web_tests/virtual/omt-worker-fetch/external/wpt/resource-timing/resource_dedicated_worker-expected.txt

This is a testharness.js-based test.
FAIL There should be six entries: 4 scripts, 1 stylesheet, and the worker itself assert_equals: There should be six entries: 4 scripts, 1 stylesheet, and the worker itself expected 6 but got 5
Harness: the test ran to completion.
